+++ content/components/accountCard.less
@@ -15,9 +15,7 @@
 /* along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-@import url('https://fonts.googleapis.com/css?family=Roboto:300');
-
-:local(.container) {
+:local(.card), :local(.tokens) {
   display: inline-flex;
   flex-direction: row;
   align-items: center;
@@ -25,4 +23,52 @@
   transition-property: opacity, transform, z-index;
   transition-duration: 0.25s;
   transition-timing-function: cubic-bezier(0.7,0,0.3,1);
+}
+
+:local(.card) {
+  position: absolute;
+  font-size: 1rem;
+
+  opacity: 0;
+  transform: scale(0.25) translateY(-185%);
+  z-index: -1;
+
+  flex-direction: column;
+  align-items: flex-start;
+
+  background-color: rgba(40, 40, 40, 0.9);
+  color: rgb(200, 200, 200);
+
+  padding: 0.75em 1em;
+
+  :local(.address) {
+    cursor: text;
+  }
+}
+
+:local(.card):local(.open) {
+  opacity: 1;
+  transform: scale(1) translateY(-50%);
+  z-index: 9999;
+}
+
+:local(.header) {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  margin-bottom: 0.5em;
+
+  :local(.title) {
+    margin-left: 1em;
+    font-size: 1.15em;
+
+    span {
+      display: block;
+      margin: 0.25em 0;
+    }
+
+    :local(.name) {
+      font-size: 0.75em;
+    }
+  }
 }
