+++ content/components/augmentedIcon.less
@@ -15,9 +15,7 @@
 /* along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-@import url('https://fonts.googleapis.com/css?family=Roboto:300');
-
-:local(.container) {
+:local(.badges), :local(.icons) {
   display: inline-flex;
   flex-direction: row;
   align-items: center;
@@ -25,4 +23,47 @@
   transition-property: opacity, transform, z-index;
   transition-duration: 0.25s;
   transition-timing-function: cubic-bezier(0.7,0,0.3,1);
+}
+
+:local(.icons) {
+  flex-direction: column;
+  margin-left: 5px;
+  font-family: 'Roboto', sans-serif;
+  font-weight: 300;
+
+  &:hover {
+    cursor: pointer;
+
+    :local(.badges) {
+      transform: scale(1.5) translateY(-75%);
+      opacity: 1;
+      z-index: 999;
+    }
+  }
+}
+
+:local(.icons):local(.expanded) {
+  &:hover :local(.badges) {
+    transform: scale(1) translateY(-50%);
+    opacity: 0;
+    z-index: -1;
+  }
+}
+
+:local(.badges) {
+  box-sizing: content-box;
+  padding: 0 2px;
+  position: absolute;
+
+  opacity: 0;
+  transform: scale(1) translateY(-50%);
+  z-index:-1;
+
+  > * {
+    margin: 3px;
+  }
+
+  &:local(.fromBottom) {
+    transform: scale(1) translateY(-200%);
+  }
 }
