+++ content/components/badge.less
@@ -15,14 +15,7 @@
 /* along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-@import url('https://fonts.googleapis.com/css?family=Roboto:300');
-
-:local(.container) {
-  display: inline-flex;
-  flex-direction: row;
-  align-items: center;
-
-  transition-property: opacity, transform, z-index;
-  transition-duration: 0.25s;
-  transition-timing-function: cubic-bezier(0.7,0,0.3,1);
+:local(.badge) {
+  width: auto;
+  border-radius: 50%;
 }
